merge master

+++ salt/modules/upstart.py
@@ -3,6 +3,12 @@
 Module for the management of upstart systems. The Upstart system only supports
 service starting, stopping and restarting.
 
+.. important::
+    If you feel that Salt should be using this module to manage services on a
+    minion, and it is using a different module (or gives an error similar to
+    *'service.start' is not available*), see :ref:`here
+    <module-provider-override>`.
+
 Currently (as of Ubuntu 12.04) there is no tool available to disable
 Upstart services (like update-rc.d). This[1] is the recommended way to
 disable an Upstart service. So we assume that all Upstart services
@@ -43,9 +49,13 @@ from __future__ import absolute_import
 # Import python libs
 import glob
 import os
+import re
+import itertools
+import fnmatch
 
 # Import salt libs
 import salt.utils
+import salt.utils.files
 import salt.modules.cmdmod
 import salt.utils.systemd
 
@@ -63,7 +73,7 @@ def __virtual__():
     '''
     # Disable on these platforms, specific service modules exist:
     if salt.utils.systemd.booted(__context__):
-        return False
+        return (False, 'The upstart execution module failed to load: this system was booted with systemd.')
     elif __grains__['os'] in ('Ubuntu', 'Linaro', 'elementary OS', 'Mint'):
         return __virtualname__
     elif __grains__['os'] in ('Debian', 'Raspbian'):
@@ -72,7 +82,8 @@ def __virtual__():
             initctl_version = salt.modules.cmdmod._run_quiet(debian_initctl + ' version')
             if 'upstart' in initctl_version:
                 return __virtualname__
-    return False
+    return (False, 'The upstart execution module failed to load: '
+        ' the system must be Ubuntu-based, or Debian-based with upstart support.')
 
 
 def _find_utmp():
@@ -99,7 +110,7 @@ def _default_runlevel():
     # Try to get the "main" default.  If this fails, throw up our
     # hands and just guess "2", because things are horribly broken
     try:
-        with salt.utils.fopen('/etc/init/rc-sysinit.conf') as fp_:
+        with salt.utils.files.fopen('/etc/init/rc-sysinit.conf') as fp_:
             for line in fp_:
                 if line.startswith('env DEFAULT_RUNLEVEL'):
                     runlevel = line.split('=')[-1].strip()
@@ -108,7 +119,7 @@ def _default_runlevel():
 
     # Look for an optional "legacy" override in /etc/inittab
     try:
-        with salt.utils.fopen('/etc/inittab') as fp_:
+        with salt.utils.files.fopen('/etc/inittab') as fp_:
             for line in fp_:
                 if not line.startswith('#') and 'initdefault' in line:
                     runlevel = line.split(':')[1]
@@ -120,7 +131,7 @@ def _default_runlevel():
     try:
         valid_strings = set(
             ('0', '1', '2', '3', '4', '5', '6', 's', 'S', '-s', 'single'))
-        with salt.utils.fopen('/proc/cmdline') as fp_:
+        with salt.utils.files.fopen('/proc/cmdline') as fp_:
             for line in fp_:
                 for arg in line.strip().split():
                     if arg in valid_strings:
@@ -170,7 +181,12 @@ def _upstart_is_disabled(name):
     NOTE: An Upstart service can also be disabled by placing "manual"
     in /etc/init/[name].conf.
     '''
-    return os.access('/etc/init/{0}.override'.format(name), os.R_OK)
+    files = ['/etc/init/{0}.conf'.format(name), '/etc/init/{0}.override'.format(name)]
+    for file_name in itertools.ifilter(os.path.isfile, files):
+        with salt.utils.files.fopen(file_name) as fp_:
+            if re.search(r'^\s*manual', fp_.read(), re.MULTILINE):
+                return True
+    return False
 
 
 def _upstart_is_enabled(name):
@@ -219,11 +235,24 @@ def _iter_service_names():
         name = os.path.basename(line)
         found.add(name)
         yield name
-    for line in glob.glob('/etc/init/*.conf'):
-        name = os.path.basename(line)[:-5]
-        if name in found:
-            continue
-        yield name
+
+    # This walk method supports nested services as per the init man page
+    # definition 'For example a configuration file /etc/init/rc-sysinit.conf
+    # is named rc-sysinit, while a configuration file /etc/init/net/apache.conf
+    # is named net/apache'
+    init_root = '/etc/init/'
+    for root, dirnames, filenames in os.walk(init_root):
+        relpath = os.path.relpath(root, init_root)
+        for filename in fnmatch.filter(filenames, '*.conf'):
+            if relpath == '.':
+                # service is defined in the root, no need to append prefix.
+                name = filename[:-5]
+            else:
+                # service is nested, append its relative path prefix.
+                name = os.path.join(relpath, filename[:-5])
+            if name in found:
+                continue
+            yield name
 
 
 def get_enabled():
@@ -398,28 +427,53 @@ def force_reload(name):
 
 def status(name, sig=None):
     '''
-    Return the status for a service, returns a bool whether the service is
-    running.
+    Return the status for a service.
+    If the name contains globbing, a dict mapping service name to True/False
+    values is returned.
+
+    .. versionchanged:: Oxygen
+        The service name can now be a glob (e.g. ``salt*``)
+
+    Args:
+        name (str): The name of the service to check
+        sig (str): Signature to use to find the service via ps
+
+    Returns:
+        bool: True if running, False otherwise
+        dict: Maps service name to True if running, False otherwise
 
     CLI Example:
 
     .. code-block:: bash
 
-        salt '*' service.status <service name>
+        salt '*' service.status <service name> [service signature]
     '''
     if sig:
         return bool(__salt__['status.pid'](sig))
-    cmd = ['service', name, 'status']
-    if _service_is_upstart(name):
-        # decide result base on cmd output, thus ignore retcode,
-        # which makes cmd output not at error lvl even when cmd fail.
-        return 'start/running' in __salt__['cmd.run'](cmd, python_shell=False,
-                                                      ignore_retcode=True)
-    # decide result base on retcode, thus ignore output (set quite)
-    # because there is no way to avoid logging at error lvl when
-    # service is not running - retcode != 0 (which is totally relevant).
-    return not bool(__salt__['cmd.retcode'](cmd, python_shell=False,
-                                            quite=True))
+
+    contains_globbing = bool(re.search(r'\*|\?|\[.+\]', name))
+    if contains_globbing:
+        services = fnmatch.filter(get_all(), name)
+    else:
+        services = [name]
+    results = {}
+    for service in services:
+        cmd = ['service', service, 'status']
+        if _service_is_upstart(service):
+            # decide result base on cmd output, thus ignore retcode,
+            # which makes cmd output not at error lvl even when cmd fail.
+            results[service] = 'start/running' in __salt__['cmd.run'](cmd, python_shell=False,
+                                                                      ignore_retcode=True)
+        else:
+            # decide result base on retcode, thus ignore output (set quite)
+            # because there is no way to avoid logging at error lvl when
+            # service is not running - retcode != 0 (which is totally relevant).
+            results[service] = not bool(__salt__['cmd.retcode'](cmd, python_shell=False,
+                                                                ignore_retcode=True,
+                                                                quite=True))
+    if contains_globbing:
+        return results
+    return results[name]
 
 
 def _get_service_exec():
@@ -436,9 +490,11 @@ def _upstart_disable(name):
     '''
     Disable an Upstart service.
     '''
+    if _upstart_is_disabled(name):
+        return _upstart_is_disabled(name)
     override = '/etc/init/{0}.override'.format(name)
-    with salt.utils.fopen(override, 'w') as ofile:
-        ofile.write('manual')
+    with salt.utils.files.fopen(override, 'a') as ofile:
+        ofile.write('manual\n')
     return _upstart_is_disabled(name)
 
 
@@ -446,8 +502,17 @@ def _upstart_enable(name):
     '''
     Enable an Upstart service.
     '''
+    if _upstart_is_enabled(name):
+        return _upstart_is_enabled(name)
     override = '/etc/init/{0}.override'.format(name)
-    if os.access(override, os.R_OK):
+    files = ['/etc/init/{0}.conf'.format(name), override]
+    for file_name in itertools.ifilter(os.path.isfile, files):
+        with salt.utils.files.fopen(file_name, 'r+') as fp_:
+            new_text = re.sub(r'^\s*manual\n?', '', fp_.read(), 0, re.MULTILINE)
+            fp_.seek(0)
+            fp_.write(new_text)
+            fp_.truncate()
+    if os.access(override, os.R_OK) and os.path.getsize(override) == 0:
         os.unlink(override)
     return _upstart_is_enabled(name)
